fix compiler warnings, allow FIFO when cross-compiling, make sure README.logformat is shipped

+++ src/conv.c
@@ -367,7 +367,7 @@ const char *ConvToUTF8 (const char *inn, UBYTE enc)
     if (!inn)
         return "";
     
-    t = s_catf (t, &size, "%*s", strlen (in) * 3, "");
+    t = s_catf (t, &size, "%*s", (int)strlen (in) * 3, "");
     *t = '\0';
     
     for (*t = '\0'; *in; in++)
@@ -459,7 +459,7 @@ const char *ConvFromUTF8 (const char *inn, UBYTE enc)
     if (!inn)
         return "";
 
-    t = s_catf (t, &size, "%*s", strlen (in) * 3, "");
+    t = s_catf (t, &size, "%*s", (int)strlen (in) * 3, "");
     *t = '\0';
     
     for (*t = '\0'; *in; in++)
